refactor: update Perses integration and improve plugin management
- Updated dependencies in package.json to use the latest versions of Perses plugins.
- Renamed PersesWrapper to PersesPluginWrapper for clarity and consistency.
- Created a new PersesPluginRegistry to manage plugin loading.
- Moved API configurations to a new persesApi.ts file for better organization.
- Removed the old PersesWrapper implementation to streamline the codebase.
- Added a new PersesDashboardWrapper file for future dashboard integration.
