+++ embedding/vite/package.json
@@ -11,11 +11,23 @@
   },
   "dependencies": {
     "@mui/material": "^6.4.11",
-    "@perses-dev/components": "^0.51.0-beta.1",
-    "@perses-dev/dashboards": "^0.51.0-beta.1",
-    "@perses-dev/plugin-system": "^0.51.0-beta.1",
+    "@perses-dev/bar-chart-plugin": "^0.7.0",
+    "@perses-dev/components": "^0.51.0-rc.0",
+    "@perses-dev/core": "^0.51.0-rc.0",
+    "@perses-dev/dashboards": "^0.51.0-rc.0",
+    "@perses-dev/gauge-chart-plugin": "^0.7.0",
+    "@perses-dev/markdown-plugin": "^0.7.0",
+    "@perses-dev/panels-plugin": "^0.50.3",
+    "@perses-dev/pie-chart-plugin": "^0.7.0",
+    "@perses-dev/plugin-system": "^0.51.0-rc.0",
     "@perses-dev/prometheus-plugin": "^0.51.0-beta.2",
+    "@perses-dev/scatter-chart-plugin": "^0.6.0",
+    "@perses-dev/stat-chart-plugin": "^0.7.0",
+    "@perses-dev/static-list-variable-plugin": "^0.3.0",
+    "@perses-dev/status-history-chart-plugin": "^0.7.0",
+    "@perses-dev/table-plugin": "^0.6.0",
     "@perses-dev/timeseries-chart-plugin": "^0.6.0",
+    "@perses-dev/timeseries-table-plugin": "^0.7.0",
     "@tanstack/react-query": "^5.75.2",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
